Add files via upload
The library has a parent of Media class, its children are Book and Movie.  Using the extends keyword to attach our children to the parent class we can simplify the code needed to make several different classes that all share a certain portion of properties shorter and not repeat ourselves.  Practice using super to call from the parent class.